Remove crypto import from JS bundle
This PR removes the crypto import from the ristretto255.min.js bundle.
The import was added by rollup because rollup detected an import in the
tweetnacl source code. However, tweetnacl imports crypto as a sort of
environment check, and both browsers and recent versions of Node.js have
a crypto global with getRandomValues. In either case it's safe to use
the crypto global without trying to import.